type score int to numeric

+++ migrations/20220323132227-create-user-movie.js
@@ -13,7 +13,7 @@ module.exports = {
         allowNull: false,
       },
       score: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.NUMERIC,
         allowNull: false,
       },
       MovieId: {
